addresed the review

+++ calculator.cpp
@@ -42,7 +42,7 @@ int multiply(int a, int b) {
  return a * b;
 }
 
-int divide(int a, int b) {
+double divide(double a, double b) {
     if (b == 0) {
     cout << "Error: Division by zero!" << endl;
     return 0;
